project-files/requirements.txt
docker==3.7.0
fake-useragent==0.1.11
h5py==2.9.0
keras==2.2.4
matplotlib==3.0.2
numpy==1.15.4
pandas==0.23.4
pytest==4.2.0
pyyaml==3.13
savoir==1.0.6
scapy==2.4.2
schedule==0.6.0
scipy==1.1.0
sklearn==0.0
statsmodels==0.9.0
tensorflow==1.12.0
theano==1.0.3
virtualenv==16.3.0
xlrd==0.9.0
